+++ apps/web-giddh/src/app/shared/tax-sidebar/tax-sidebar.component.html
@@ -160,7 +160,7 @@
                             [routerLink]="['/pages/vat-report/liability-report']"
                             [routerLinkActiveOptions]="{exact: true}"
                         >
-                            <span><i class="icon-vat1"></i></span>
+                            <span><i class="icon-balance-sheet-new"></i></span>
                             <span>{{ localeData?.liability_report }}</span>
                         </a>
                     </li>
