+++ apps/web-giddh/src/app/vat-report/liability-detailed-report/liability-detailed-report.component.html
@@ -96,12 +96,20 @@
                 </div>
             </section>
         </div>
-        <div class="pagination-wrapper" *ngIf="vatLiabilityDetailedReport.totalPages > 1">
-            <pagination [totalItems]="vatReportTransactions.totalPages" [(ngModel)]="vatReportTransactions.page"
-                (pageChanged)="pageChanged($event)" [maxSize]="6" [boundaryLinks]="true" [itemsPerPage]="1"
-                [rotate]="false" [firstText]="commonLocaleData?.app_first"
-                [previousText]="commonLocaleData?.app_previous" [nextText]="commonLocaleData?.app_next"
-                [lastText]="commonLocaleData?.app_last" class="pagination-sm"></pagination>
+        <div class="pagination-wrapper" *ngIf="vatLiabilityDetailedReport?.totalPages > 1">
+            <pagination 
+                [totalItems]="vatLiabilityDetailedReport?.totalItems" 
+                [(ngModel)]="vatLiabilityDetailedReport.page"
+                (pageChanged)="pageChanged($event)" 
+                [maxSize]="vatLiabilityDetailedReport?.totalPages >= 5 ? 5 : vatLiabilityDetailedReport?.totalPages" 
+                [boundaryLinks]="true" 
+                [itemsPerPage]="vatLiabilityReportRequest.count"
+                [rotate]="false" 
+                [firstText]="commonLocaleData?.app_first"
+                [previousText]="commonLocaleData?.app_previous" 
+                [nextText]="commonLocaleData?.app_next"
+                [lastText]="commonLocaleData?.app_last" class="pagination-sm"
+            ></pagination>
         </div>
     </div>
 </ng-container>
