+++ apps/web-giddh/src/app/vat-report/vat-report-filters/vat-report-filters.component.html
@@ -1,0 +1,66 @@
+<div class="advance-search-by-date date-advance-search justify-content-center row-gap">
+    <dropdown-field [options]="financialYears" (selectedOption)="getYearStartAndEndDate($event)"
+        [placeholder]="localeData?.select_year_label" [defaultValue]="selectedYear"></dropdown-field>
+
+    <dropdown-field [options]="months" *ngIf="year && selectedYear" (selectedOption)="getMonthStartAndEndDate($event)"
+        [placeholder]="localeData?.select_month_label" [defaultValue]="selectedMonth?.label"></dropdown-field>
+    <div>
+        <div class="input-group custom-datepicker custom-new-datepicker" (click)="showGiddhDatepicker($event)">
+            <input type="text" name="selectedDateRange" [value]="selectedDateRangeUi ? selectedDateRangeUi : ''"
+                class="giddh-datepicker date-range-picker border-radius-2" />
+        </div>
+    </div>
+    <div *ngIf="
+            currentCompanyBranches?.length > 2 && currentOrganizationType === 'COMPANY'
+        " class="input-width">
+        <dropdown-field [label]="commonLocaleData?.app_entity" [options]="currentCompanyBranches" width="100%"
+            (selectedOption)="
+                currentBranch.uniqueName = $event.value; handleBranchChange($event)
+            " [placeholder]="localeData?.select_entity"></dropdown-field>
+    </div>
+    <div *ngIf="taxes?.length > 1" class="input-width">
+        <dropdown-field
+            [label]="isKenyaCompany || isZimbabweCompany ? commonLocaleData?.app_vat : commonLocaleData?.app_trn"
+            [options]="taxes" (selectedOption)="getVatReport($event)" [placeholder]="commonLocaleData?.select_trn"
+            [defaultValue]="taxes[0]?.label" [required]="true">
+        </dropdown-field>
+    </div>
+
+    <div class="ml-auto d-flex justify-content-center column-gap1 align-items-center">
+        <div *ngIf="isZimbabweCompany && taxNumber" class="currency-width">
+            <dropdown-field [label]="commonLocaleData?.app_currency" [options]="vatReportCurrencyList"
+                (selectedOption)="onCurrencyChange($event)" [placeholder]="commonLocaleData?.app_select"
+                [defaultValue]="vatReportCurrencyCode" [required]="true">
+            </dropdown-field>
+        </div>
+        <div *ngIf="taxNumber && (isZimbabweCompany || isKenyaCompany)" class="export-btn w100">
+            <button mat-stroked-button color="primary" (click)="downloadVatReport()">
+                <i class="icon-download1"></i>
+                {{ commonLocaleData?.app_export }}
+            </button>
+        </div>
+        <a mat-stroked-button color="primary" class="text-right" *ngIf="isUKCompany" [href]="connectToHMRCUrl"
+            [disabled]="!connectToHMRCUrl">
+            <ng-container *ngIf="connectToHMRCUrl">
+                {{ localeData?.connect_to_hmrc }}
+            </ng-container>
+            <ng-container *ngIf="!connectToHMRCUrl">
+                {{ localeData?.connected_to_hmrc }}
+            </ng-container>
+        </a>
+    </div>
+</div>
+
+<ng-template #datepickerTemplate>
+    <div class="datepicker-modal">
+        <div class="modal-body">
+            <app-datepicker-wrapper [inputStartDate]="selectedDateRange?.startDate"
+                [inputEndDate]="selectedDateRange?.endDate" [alwaysShowCalendars]="true" [ranges]="datePickerOption"
+                [selectedRangeLabel]="selectedRangeLabel" [showCustomRangeLabel]="true" [showClearButton]="false"
+                [showCancel]="true" [linkedCalendars]="true" [showDropdowns]="true"
+                (rangeClicked)="dateSelectedCallback($event)" (datesUpdated)="dateSelectedCallback($event)"
+                [keepCalendarOpeningWithRange]="false" [showRangeLabelOnInput]="false"
+                [dateFieldPosition]="dateFieldPosition"></app-datepicker-wrapper>
+        </div>
+    </div>
+</ng-template>
