+++ apps/web-giddh/src/app/vat-report/vat-report-filters/vat-report-filters.component.scss
@@ -1,0 +1,37 @@
+.advance-search-by-date {
+    display: flex;
+    column-gap: 15px;
+
+    .input-wrapper {
+        display: flex;
+
+        .btn-group {
+            button {
+                margin: 0 !important;
+                height: 100%;
+            }
+        }
+    }
+
+    .input-width {
+        min-width: 220px;
+    }
+
+    .currency-width {
+        min-width: 100px;
+    }
+
+    .custom-datepicker {
+        &.custom-new-datepicker {
+            input {
+                width: 240px !important;
+                max-width: 240px;
+            }
+        }
+    }
+
+    .export-btn {
+        display: flex;
+        justify-content: flex-end;
+    }
+}
