+++ apps/web-giddh/src/app/vat-report/vat-report.component.ts
@@ -1,4 +1,4 @@
-import { Observable, ReplaySubject } from 'rxjs';
+import { ReplaySubject } from 'rxjs';
 import { takeUntil } from 'rxjs/operators';
 import { ChangeDetectorRef, Component, TemplateRef, OnDestroy, OnInit, ViewChild, } from '@angular/core';
 import { Router } from '@angular/router';
@@ -9,16 +9,8 @@ import { GeneralService } from '../services/general.service';
 import { ToasterService } from '../services/toaster.service';
 import { VatService } from "../services/vat.service";
 import * as dayjs from 'dayjs';
-import { GIDDH_DATE_FORMAT, GIDDH_NEW_DATE_FORMAT_UI } from "../shared/helpers/defaultDateFormat";
 import { saveAs } from "file-saver";
-import { IOption } from '../theme/ng-select/ng-select';
-import { GstReconcileService } from '../services/gst-reconcile.service';
-import { SettingsBranchActions } from '../actions/settings/branch/settings.branch.action';
-import { OrganizationType } from '../models/user-login-state';
 import { BreakpointObserver, BreakpointState } from '@angular/cdk/layout';
-import { cloneDeep } from '../lodash-optimized';
-import { BsModalRef, BsModalService } from 'ngx-bootstrap/modal';
-import { GIDDH_DATE_RANGE_PICKER_RANGES } from '../app.constant';
 import { SettingsFinancialYearService } from '../services/settings.financial-year.service';
 @Component({
     selector: 'app-vat-report',
@@ -28,55 +20,23 @@ import { SettingsFinancialYearService } from '../services/settings.financial-yea
 export class VatReportComponent implements OnInit, OnDestroy {
     public vatReport: any[] = [];
     public activeCompany: any;
-    /** This will store available date ranges */
-    public datePickerOption: any = GIDDH_DATE_RANGE_PICKER_RANGES;
-    public dayjs = dayjs;
     public fromDate: string = '';
     public toDate: string = '';
-    /** Hold selected year */
-    public year: number = null;
     private destroyed$: ReplaySubject<boolean> = new ReplaySubject(1);
-    @ViewChild('periodDropdown', { static: true }) public periodDropdown;
-    /** directive to get reference of element */
-    @ViewChild('datepickerTemplate') public datepickerTemplate: TemplateRef<any>;
-    public isMonthSelected: boolean = true;
-    public currentPeriod: any = {};
-    public showCalendar: boolean = false;
-    public datepickerVisibility: any = 'hidden';
-    /** Stores the tax details of a company */
-    public taxes: IOption[] = [];
     /** Tax number */
     public taxNumber: string;
     /** True, if API is in progress */
     public isTaxApiInProgress: boolean;
-    /** This holds giddh date format */
-    public giddhDateFormat: string = GIDDH_DATE_FORMAT;
-    /** Observable to store the branches of current company */
-    public currentCompanyBranches$: Observable<any>;
-    /** Stores the branch list of a company */
-    public currentCompanyBranches: Array<any>;
     /** Stores the current branch */
     public currentBranch: any = { name: '', uniqueName: '' };
     /** This will hold local JSON data */
     public localeData: any = {};
     /** This will hold common JSON data */
     public commonLocaleData: any = {};
-    /** Stores the current organization type */
-    public currentOrganizationType: OrganizationType;
     /** This will hold the value out/in to open/close setting sidebar popup */
     public asideGstSidebarMenuState: string = 'in';
     /** this will check mobile screen size */
     public isMobileScreen: boolean = false;
-    /** This will store the x/y position of the field to show datepicker under it */
-    public dateFieldPosition: any = { x: 0, y: 0 };
-    /** Datepicker modal reference */
-    public modalRef: BsModalRef;
-    /** Selected range label */
-    public selectedRangeLabel: any = "";
-    /* This will store selected date range to use in api */
-    public selectedDateRange: any;
-    /** This will store selected date range to show on UI */
-    public selectedDateRangeUi: any;
     /** Hold uae main table displayed columns */
     public displayedColumns: string[] = ['number', 'name', 'aed_amt', 'vat_amt', 'adjustment'];
     /** Hold uae bottom table displayed columns */
@@ -101,63 +61,27 @@ export class VatReportComponent implements OnInit, OnDestroy {
     public isZimbabweCompany: boolean = false;
     /** True if active country is Kenya */
     public isKenyaCompany: boolean = false;
-    /** Hold static months array */
-    public months: any[] = [
-        { label: 'January', value: 1 },
-        { label: 'Febuary', value: 2 },
-        { label: 'March', value: 3 },
-        { label: 'April', value: 4 },
-        { label: 'May', value: 5 },
-        { label: 'June', value: 6 },
-        { label: 'July', value: 7 },
-        { label: 'August', value: 8 },
-        { label: 'September', value: 9 },
-        { label: 'October', value: 10 },
-        { label: 'November', value: 11 },
-        { label: 'December', value: 12 },
-    ];
-    /** Hold financial year list */
-    public financialYears: any = [];
     /** True if api call in progress */
     public isLoading: boolean = false;
-    /** Hold selected month value */
-    public selectedMonth: any = "";
-    /** Hold selected year value */
-    public selectedYear: any = "";
     /** Hold HMRC portal url */
     public connectToHMRCUrl: string = null;
-    /** Holds Currency List for Zimbabwe Amount exchange rate */
-    public vatReportCurrencyList: any[] = [
-        { label: 'BWP', value: 'BWP', additional: { symbol: 'P' } },
-        { label: 'USD', value: 'USD', additional: { symbol: '$' } },
-        { label: 'GBP', value: 'GBP', additional: { symbol: '£' } },
-        { label: 'INR', value: 'INR', additional: { symbol: '₹' } },
-        { label: 'EUR', value: 'EUR', additional: { symbol: '€' } }
-    ];
     /** Holds Current Currency Code for Zimbabwe report */
-    public vatReportCurrencyCode: 'BWP' | 'USD' | 'GBP' | 'INR' | 'EUR' = this.vatReportCurrencyList[0].value;
+    public vatReportCurrencyCode: 'BWP' | 'USD' | 'GBP' | 'INR' | 'EUR' = 'BWP';
     /** Holds Current Currency Symbol for Zimbabwe report */
-    public vatReportCurrencySymbol: string = this.vatReportCurrencyList[0].additional.symbol;
+    public vatReportCurrencySymbol: string = 'P';
     /** Holds Current Currency Map Amount Decimal currency wise for Zimbabwe report */
     public vatReportCurrencyMap: string[];
-    /** True if Current branch has Tax Number */
-    public hasTaxNumber: boolean = false;
 
     constructor(
-        private gstReconcileService: GstReconcileService,
         private store: Store<AppState>,
         private vatService: VatService,
         private generalService: GeneralService,
         private toasty: ToasterService,
         private cdRef: ChangeDetectorRef,
         private route: Router,
-        private settingsBranchAction: SettingsBranchActions,
         private breakpointObserver: BreakpointObserver,
-        private modalService: BsModalService,
         public settingsFinancialYearService: SettingsFinancialYearService
-    ) {
-        this.getFinancialYears();
-    }
+    ) { }
 
     public ngOnInit() {
         this.store.pipe(select(state => state.session.activeCompany), takeUntil(this.destroyed$)).subscribe(activeCompany => {
@@ -181,6 +105,7 @@ export class VatReportComponent implements OnInit, OnDestroy {
                     this.asideGstSidebarMenuState = 'in';
                 }
             });
+            
         this.store.pipe(select(appState => appState.general.openGstSideMenu), takeUntil(this.destroyed$)).subscribe(shouldOpen => {
             if (this.isMobileScreen) {
                 if (shouldOpen) {
@@ -188,63 +113,6 @@ export class VatReportComponent implements OnInit, OnDestroy {
                 } else {
                     this.asideGstSidebarMenuState = 'out';
                 }
-            }
-        });
-
-        this.currentOrganizationType = this.generalService.currentOrganizationType;
-        this.currentCompanyBranches$ = this.store.pipe(select(appStore => appStore.settings.branches), takeUntil(this.destroyed$));
-        this.currentCompanyBranches$.subscribe(response => {
-            if (response && response.length) {
-                this.currentCompanyBranches = response.map(branch => ({
-                    label: branch.alias,
-                    value: branch?.uniqueName,
-                    name: branch.name,
-                    parentBranch: branch.parentBranch
-                }));
-                this.currentCompanyBranches.unshift({
-                    label: this.activeCompany ? this.activeCompany.nameAlias || this.activeCompany.name : '',
-                    name: this.activeCompany ? this.activeCompany.name : '',
-                    value: this.activeCompany ? this.activeCompany.uniqueName : '',
-                    isCompany: true
-                });
-                let currentBranchUniqueName;
-                if (!this.currentBranch?.uniqueName) {
-                    // Assign the current branch only when it is not selected. This check is necessary as
-                    // opening the branch switcher would reset the current selected branch as this subscription is run everytime
-                    // branches are loaded
-                    if (this.currentOrganizationType === OrganizationType.Branch) {
-                        currentBranchUniqueName = this.generalService.currentBranchUniqueName;
-                        this.currentBranch = cloneDeep(response.find(branch => branch?.uniqueName === currentBranchUniqueName));
-                        this.hasTaxNumber = this.currentBranch?.addresses?.filter(address => address?.taxNumber?.length > 0)?.length > 0;
-                    } else {
-                        currentBranchUniqueName = this.activeCompany ? this.activeCompany.uniqueName : '';
-                        this.currentBranch = {
-                            name: this.activeCompany ? this.activeCompany.name : '',
-                            alias: this.activeCompany ? this.activeCompany.nameAlias || this.activeCompany.name : '',
-                            uniqueName: this.activeCompany ? this.activeCompany.uniqueName : '',
-                        };
-                    }
-                    if (this.hasTaxNumber || this.currentOrganizationType === OrganizationType.Company) {
-                        this.loadTaxDetails();
-                    }
-                }
-            } else {
-                if (this.generalService.companyUniqueName) {
-                    // Avoid API call if new user is onboarded
-                    this.store.dispatch(this.settingsBranchAction.GetALLBranches({ from: '', to: '' }));
-                }
-            }
-        });
-        // Refresh report data according to universal date
-        this.store.pipe(select(state => state.session.applicationDate), takeUntil(this.destroyed$)).subscribe((dateObj: Date[]) => {
-            if (dateObj) {
-                this.selectedDateRange = { startDate: dayjs(dateObj[0]), endDate: dayjs(dateObj[1]) };
-                this.selectedDateRangeUi = dayjs(dateObj[0]).format(GIDDH_NEW_DATE_FORMAT_UI) + " - " + dayjs(dateObj[1]).format(GIDDH_NEW_DATE_FORMAT_UI);
-                this.fromDate = dayjs(dateObj[0]).format(GIDDH_DATE_FORMAT);
-                this.toDate = dayjs(dateObj[1]).format(GIDDH_DATE_FORMAT);
-                this.selectedMonth = "";
-                this.selectedYear = "";
-                this.year = null;
             }
         });
     }
@@ -259,14 +127,9 @@ export class VatReportComponent implements OnInit, OnDestroy {
     /**
      * This will use for get vat report for uae and uk according to country code
      *
-     * @param {*} [event]
      * @memberof VatReportComponent
      */
-    public getVatReport(event?: any): void {
-        if (event && event.value) {
-            this.taxNumber = event.value;
-        }
-
+    public getVatReport(): void {
         if (this.taxNumber) {
             let countryCode;
             let vatReportRequest = new VatReportRequest();
@@ -286,6 +149,7 @@ export class VatReportComponent implements OnInit, OnDestroy {
             }
             this.vatReport = [];
             this.isLoading = true;
+
             if (!this.isUKCompany && !this.isZimbabweCompany && !this.isKenyaCompany) {
                 this.vatService.getVatReport(vatReportRequest).pipe(takeUntil(this.destroyed$)).subscribe((res) => {
                     if (res) {
@@ -305,8 +169,6 @@ export class VatReportComponent implements OnInit, OnDestroy {
                         if (res && res?.status === 'success' && res?.body) {
                             this.vatReport = res.body?.sections;
                             if (this.isZimbabweCompany) {
-                                this.vatReportCurrencyCode = res.body?.currency?.code;
-                                this.vatReportCurrencySymbol = this.vatReportCurrencyList.filter(item => item.label === res.body?.currency?.code).map(item => item.additional.symbol).join();
                                 this.vatReportCurrencyMap = res.body?.currencyMap;
                             }
 
@@ -350,45 +212,6 @@ export class VatReportComponent implements OnInit, OnDestroy {
     }
 
     /**
-     * This will use for get month start date and end date
-     *
-     * @param {number} selectedMonth
-     * @memberof VatReportComponent
-     */
-    public getMonthStartAndEndDate(selectedMonth: any) {
-        if (selectedMonth) {
-            this.selectedMonth = selectedMonth;
-
-            // Month is zero-based, so subtract 1 from the selected month
-            const startDate = new Date(this.year, selectedMonth.value - 1, 1);
-            const endDate = new Date(this.year, selectedMonth.value, 0);
-            this.selectedDateRange = { startDate: dayjs(startDate), endDate: dayjs(endDate) };
-            this.selectedDateRangeUi = dayjs(startDate).format(GIDDH_NEW_DATE_FORMAT_UI) + " - " + dayjs(endDate).format(GIDDH_NEW_DATE_FORMAT_UI);
-            this.fromDate = dayjs(startDate).format(GIDDH_DATE_FORMAT);
-            this.toDate = dayjs(endDate).format(GIDDH_DATE_FORMAT);
-            this.getVatReport();
-
-        }
-    }
-
-    /**
-    * This will use for get year selected
-    *
-    * @param {*} selectedYear
-    * @memberof VatReportComponent
-    */
-    public getYearStartAndEndDate(selectedYear: any): void {
-        if (selectedYear?.value && selectedYear?.label !== this.selectedYear) {
-            this.year = Number(selectedYear?.value);
-            this.selectedYear = selectedYear.label;
-
-            if (this.selectedMonth) {
-                this.getMonthStartAndEndDate(this.selectedMonth);
-            }
-        }
-    }
-
-    /**
     * This will redirect to vat report detail page
     *
     * @param {*} section
@@ -399,97 +222,12 @@ export class VatReportComponent implements OnInit, OnDestroy {
     }
 
     /**
-     * Branch change handler
-     *
-     * @memberof VatReportComponent
-     */
-    public handleBranchChange(selectedEntity: any): void {
-        this.currentBranch.name = selectedEntity.label;
-        this.getVatReport();
-    }
-
-    /**
-     * Loads the tax details of a company
-     *
-     * @private
-     * @memberof VatReportComponent
-     */
-    private loadTaxDetails(): void {
-        this.isTaxApiInProgress = true;
-        this.gstReconcileService.getTaxDetails().pipe(takeUntil(this.destroyed$)).subscribe(response => {
-            if (response && response.body) {
-                this.taxes = response.body.map(tax => ({
-                    label: tax,
-                    value: tax
-                }));
-            }
-            this.isTaxApiInProgress = false;
-            this.taxNumber = this.taxes[0]?.value;
-            setTimeout(() => {
-                this.getVatReport();
-            },100);
-        });
-    }
-
-    /**
      * Handles GST Sidebar Navigation
      *
      * @memberof VatReportComponent
      */
     public handleNavigation(): void {
         this.route.navigate(['pages', 'gstfiling']);
-    }
-    /**
-     * To show the datepicker
-     *
-     * @param {*} element
-     * @memberof VatReportComponent
-     */
-    public showGiddhDatepicker(element: any): void {
-        if (element) {
-            this.dateFieldPosition = this.generalService.getPosition(element.target);
-        }
-        this.modalRef = this.modalService.show(
-            this.datepickerTemplate,
-            Object.assign({}, { class: 'modal-lg giddh-datepicker-modal', backdrop: false, ignoreBackdropClick: false })
-        );
-    }
-    /**
-     * This will hide the datepicker
-     *
-     * @memberof VatReportComponent
-     */
-    public hideGiddhDatepicker(): void {
-        this.modalRef.hide();
-    }
-
-    /**
-    * Call back function for date/range selection in datepicker
-    *
-    * @param {*} value
-    * @memberof VatReportComponent
-    */
-    public dateSelectedCallback(value?: any): void {
-        if (value && value.event === "cancel") {
-            this.hideGiddhDatepicker();
-            return;
-        }
-        this.selectedRangeLabel = "";
-        this.selectedMonth = "";
-        this.selectedYear = "";
-        this.year = null;
-
-        if (value && value.name) {
-            this.selectedRangeLabel = value.name;
-        }
-        this.hideGiddhDatepicker();
-        if (value && value.startDate && value.endDate) {
-            this.selectedDateRange = { startDate: dayjs(value.startDate), endDate: dayjs(value.endDate) };
-            this.selectedDateRangeUi = dayjs(value.startDate).format(GIDDH_NEW_DATE_FORMAT_UI) + " - " + dayjs(value.endDate).format(GIDDH_NEW_DATE_FORMAT_UI);
-            this.fromDate = dayjs(value.startDate).format(GIDDH_DATE_FORMAT);
-            this.toDate = dayjs(value.endDate).format(GIDDH_DATE_FORMAT);
-            this.getVatReport();
-        }
     }
 
     /**
@@ -503,56 +241,5 @@ export class VatReportComponent implements OnInit, OnDestroy {
                 this.connectToHMRCUrl = res?.body;
             }
         })
-    }
-
-    /**
-     * This will get all the financial years of the company
-     *
-     * @memberof VatReportComponent
-     */
-    private getFinancialYears(): void {
-        this.settingsFinancialYearService.getFinancialYearLimits().pipe(takeUntil(this.destroyed$)).subscribe(res => {
-            if (res.body.startDate && res.body.endDate) {
-                this.createFinancialYearsList(res.body.startDate, res.body.endDate)
-            }
-        });
-    }
-
-    /**
-    * Create financial years list for dropdown
-    *
-    * @private
-    * @param {*} startDate
-    * @param {*} endDate
-    * @memberof VatReportComponent
-    */
-    private createFinancialYearsList(startDate: any, endDate: any): any {
-        if (startDate && endDate) {
-            let startYear = startDate.split('-');
-            startYear = startYear[startYear?.length - 1];
-
-            let endYear = endDate.split('-');
-            endYear = endYear[endYear?.length - 1];
-
-            this.financialYears = [
-                { label: startYear, value: startYear },
-                { label: endYear, value: endYear }
-            ];
-
-            return { startYear: startYear, endYear: endYear };
-        }
-    }
-
-    /**
-     * Handle Currency change dropdown and call VAT Report API
-     *
-     * @param {*} event
-     * @memberof VatReportComponent
-     */
-    public onCurrencyChange(event: any): void {
-        if (event) {
-            this.vatReportCurrencyCode = event.value;
-            this.getVatReport();
-        }
     }
 }
